+++ src/api/axios.ts
@@ -16,30 +16,3 @@ axiosInstance.interceptors.request.use((config) => {
   }
   return config
 })
-
-// 스터디 공고 목록 조회
-export const getRecruitments = async (params: {
-  search?: string
-  category?: string
-  sort?: string
-}) => {
-  const queryParams = new URLSearchParams()
-  if (params.search) queryParams.append('search', params.search)
-  if (params.category) queryParams.append('category', params.category)
-  if (params.sort) queryParams.append('sort', params.sort)
-
-  const response = await axiosInstance.get(
-    `/recruitments?${queryParams.toString()}`
-  )
-  return response.data
-}
-
-// 스터디 공고 상세 조회
-export const getRecruitmentDetail = async (id: string) => {
-  const res = await axiosInstance.get(`/recruitments/${id}`)
-  return res.data
-}
-
-export const incrementRecruitmentViews = async (id: string) => {
-  await axiosInstance.post(`/recruitments/${id}/views`)
-}
