+++ src/api/recruitments.ts
@@ -1,0 +1,26 @@
+import { axiosInstance } from '@/api/axios'
+
+export const getRecruitments = async (params: {
+  search?: string
+  category?: string
+  sort?: string
+}) => {
+  const queryParams = new URLSearchParams()
+  if (params.search) queryParams.append('search', params.search)
+  if (params.category) queryParams.append('category', params.category)
+  if (params.sort) queryParams.append('sort', params.sort)
+
+  const response = await axiosInstance.get(
+    `/v1/recruitments?${queryParams.toString()}`
+  )
+  return response.data
+}
+
+export const getRecruitmentDetail = async (id: string) => {
+  const res = await axiosInstance.get(`/v1/recruitments/${id}`)
+  return res.data
+}
+
+export const incrementRecruitmentViews = async (id: string) => {
+  await axiosInstance.post(`/v1/recruitments/${id}/views`)
+}
